requirements.txt
pandas==2.0.3
tqdm==4.66.1
tensorboard==2.14.0
torch==2.1.2
torchvision==0.16.2
hydra-core==1.3.2
omegaconf==2.3.0
scikit-learn==1.3.2
opencv-python==4.9.0.80
protobuf==4.25
numpy~=1.26.4
pillow~=10.2.0
PyYAML~=6.0.1
jupyter~=1.0.0
pip~=24.1.1
attrs~=23.2.0
argparse~=1.4.0
wheel~=0.43.0
tornado~=6.4
Jinja2~=3.1.3
filelock~=3.13.1
cffi~=1.16.0
defusedxml~=0.7.1
typing_extensions~=4.9.0
ipython~=8.20.0
rsa~=4.9
pyasn1~=0.6.0
pytz~=2024.1
requests~=2.31.0
rich~=13.7.1
keras~=3.2.1
colorama~=0.4.6
ipywidgets~=8.1.1
notebook~=7.0.7
pytest~=8.2.0
packaging~=23.2
setuptools~=58.1.0
overrides~=7.6.0
scipy~=1.12.0
threadpoolctl~=3.3.0
matplotlib~=3.8.3
mpmath~=1.3.0
sympy~=1.12
platformdirs~=4.1.0
decorator~=5.1.1
astunparse~=1.6.3
safetensors~=0.4.3
transformers~=4.40.1
dill~=0.3.8
tensorflow~=2.16.1
psutil~=5.9.8
urllib3~=2.1.0
fsspec~=2024.3.1
aiohttp~=3.9.3
yarl~=1.9.4
pyasn1_modules~=0.4.0
cachetools~=5.3.3
certifi~=2023.11.17
MarkupSafe~=2.1.4
joblib~=1.3.2
SQLAlchemy~=2.0.28
python-dateutil~=2.8.2
pyarrow~=15.0.0
datasets~=2.19.0
six~=1.16.0
Markdown~=3.6
Pygments~=2.17.2
networkx~=3.3
oauthlib~=3.2.2
blinker~=1.7.0
idna~=3.6
Werkzeug~=3.0.1
click~=8.1.7